Support both <Description> and <PacageDescription>.
