+++ dotnet-project-file-analyzers.sln
@@ -148,6 +148,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DoubleProjectReferences", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChooseWhen", "projects\ChooseWhen\ChooseWhen.csproj", "{12251088-DE11-4DF8-871A-AD731663D2FC}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PackageDescription", "projects\PackageDescription\PackageDescription.csproj", "{B0A4EBF1-7614-4E4D-BCF0-FC00F181728F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -366,6 +368,10 @@ Global
 		{12251088-DE11-4DF8-871A-AD731663D2FC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{12251088-DE11-4DF8-871A-AD731663D2FC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{12251088-DE11-4DF8-871A-AD731663D2FC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B0A4EBF1-7614-4E4D-BCF0-FC00F181728F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B0A4EBF1-7614-4E4D-BCF0-FC00F181728F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B0A4EBF1-7614-4E4D-BCF0-FC00F181728F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B0A4EBF1-7614-4E4D-BCF0-FC00F181728F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -426,6 +432,7 @@ Global
 		{B7FC4C41-DB01-4601-BD69-D3A6E901CBDB} = {46AE4B0F-B941-4E9F-B307-7D7D1E168F64}
 		{52DD6A7B-995B-40F7-9500-5F643064E960} = {46AE4B0F-B941-4E9F-B307-7D7D1E168F64}
 		{12251088-DE11-4DF8-871A-AD731663D2FC} = {46AE4B0F-B941-4E9F-B307-7D7D1E168F64}
+		{B0A4EBF1-7614-4E4D-BCF0-FC00F181728F} = {46AE4B0F-B941-4E9F-B307-7D7D1E168F64}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3AD1EF71-7465-4ED1-81AD-504C430C8251}
