+++ projects/PackageDescription/PackageDescription.csproj
@@ -1,0 +1,31 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net7.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <IsPackable>true</IsPackable>
+  </PropertyGroup>
+
+  <PropertyGroup Label="PackageInfo">
+    <Version>1.0.0</Version>
+    <Authors>Corniel Nobel; Wesley Baartman</Authors>
+    <PackageId>DotNetProjectFile.Anayzers</PackageId>
+    <PackageDescription>Awesome library</PackageDescription>
+    <PackageTags>analyzer; coolthings</PackageTags>
+    <RepositoryUrl>https://github.com/Corniel/dotnet-project-file-analyzers</RepositoryUrl>
+    <PackageProjectUrl>https://github.com/Corniel/dotnet-project-file-analyzers</PackageProjectUrl>
+    <Copyright>Copyright © Corniel Nobel 2023-current</Copyright>
+    <PackageReleaseNotes>
+      v1.0.0
+      - Proj0001: MS Build project file could not be located. (NEW RULE)
+      - Proj0002: Upgrade legacy MS Build project file. (NEW RULE)
+      - Proj0003: Define usings explicit. (NEW RULE)
+      - Proj1001: Use analyzers for packages. (NEW RULE)
+    </PackageReleaseNotes>
+    <PackageReadmeFile>README.md</PackageReadmeFile>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <PackageIcon>logo_128x128.png</PackageIcon>
+    <PackageIconUrl>https://raw.githubusercontent.com/Corniel/dotnet-project-file-analyzers/main/design/logo_128x128.png</PackageIconUrl>
+  </PropertyGroup>
+
+</Project>
