+++ projects/PackageDescription/Placeholder.cs
@@ -1,0 +1,4 @@
+﻿namespace PackageDescription;
+
+[System.Serializable]
+public sealed class Placeholder { }
